requirements.txt
numpy>=1.19.4
Pillow>=8.0.1

#We need scikit-image >0.18, this is currently not released as a pip package
# To install it, do the following
# git clone https://github.com/scikit-image/scikit-image.git
# cd scikit-image
# pip install -e .

# Note it does not work for python > 3.8
